requirements.txt
Django==2.0.3
djangorestframework==3.7.7
pytz==2018.3
requests==2.18.4
